requirements.txt
flask
python-dotenv
psycopg2-binary
pandas
scikit-learn
flask-cors